model/requirements.txt
ultralytics
gradio
pillow
fastapi
uvicorn
python-multipart